refactor(openfoam): split blockMeshDict rendering into block_mesh.py
The approved plan for #1575 requires the case_builder split to land before any
feature edit. case_builder.py was 420 lines with an 81-line
_write_block_mesh_dict, both over the 400/50 limits the batch modules already
enforce.

Deterministic blockMeshDict body rendering moves to block_mesh.py behind
render_block_mesh_dict_body(domain), matching the render_*_dict_body convention
already used by motion.py and partial_fill.py. The builder keeps orchestration
and still composes the FoamFile header and trailing rule, so no rendered byte
moves.

case_builder.py 420 -> 346 lines; the writer method is now 6 lines.

Verified no behaviour change: sha256 of every emitted file across all five
CaseType values (66 files) is identical before and after the split.
